refactor: change name of directory
